backend/requirements.txt
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
factory-boy==3.3.0
Faker==37.5.3
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
mdurl==0.1.2
orjson==3.11.2
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
